`rpadmin`: add `singleKeyConfig()`
`rpk cluster config get` requests only a single cluster property from
`redpanda`.

However, currently we return the entire JSON of the cluster properties
to `rpk`, and then parse the desired property from there.

To support retrieving only a single value from the cluster properties
(and also to support aliased properties in the future), add
`singleKeyConfig()`, which uses a string to be used as the query parameter
`key` to the `redpanda` admin endpoint.